resize background image so app does  not crash, and swap contents of lockerHome and register activities
