+++ app/src/main/res/layout/locker_home.xml
@@ -1,31 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-
     android:orientation="vertical" android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
-    <TextView android:text="Locker Home Screen" android:layout_width="wrap_content"
-        android:layout_height="wrap_content"  />
-
-
-
-    <ImageButton
-        android:id="@+id/speakerButton"
-        android:layout_width="fill_parent"
-        android:layout_height="wrap_content"
-        android:layout_margin="10dp"
-        android:layout_marginRight="10dp"
-        android:layout_marginTop="10dp"
-        android:src="@android:drawable/ic_btn_speak_now" />
-
-    <TextView
-        android:id="@+id/speechToText"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginLeft="10dp"
-        android:layout_marginRight="10dp"
-        android:layout_marginTop="10dp"
-        android:textAppearance="?android:attr/textAppearanceLarge" />
-
+    android:layout_height="match_parent"
+    android:id="@+id/registerLayout">
+    <TextView android:text="locker home!" android:layout_width="wrap_content"
+        android:layout_height="wrap_content"/>
 
 </LinearLayout>
